implemented data managers for values that need to be visible client side
This does not work 100% as it does not really refresh when taking out
healing potions....

+++ src/main/java/com/teamcqr/chocolatequestrepoured/objects/entity/mobs/EntityCQREnderman.java
@@ -40,8 +40,8 @@ public class EntityCQREnderman extends AbstractEntityCQR {
 	}
 	
 	@Override
-	public double getSizeVariation() {
-		return 0D;
+	public float getSizeVariation() {
+		return 0F;
 	} 
 	
 	@Override
